cleanings
cutting some unnecessary draft code. it's still giving error not being able to split the ten names group in two equal parts. sometimes it splits in 5 -5/ 6-4 / 7-3.. also not sure why when clicking the button twice it continues to add names to the teams already made..?? it should random the 5 names teams already made.